statistical and update structure
